+++ src/HomePage/Home.css
@@ -2,6 +2,8 @@
     top: 20%;
     position: absolute;
     width: 100%;
+    height: 80vh;
+    /* background-color: beige; */
 }
 
 .info_bar {
@@ -59,7 +61,8 @@
     position: absolute;
     left: 25%;
     right: 6%;
-    top: 110%;
+    /* top:1px; */
+    /* top: 110%; */
     width: 70%;
     /* bottom: 2%; */
     height: 60vh;
